Rework authentication - now handled in service worker
Notifications come in to the service worker. We want actions, such as
'pupil sent to door', but for that we need to make an authorized fetch.

Therefore, authorization is now handled by serviceworker and this is no
longer a problem.

Also, we've learned a thing or two - and that's always a good thing!